Fix eventsListener and  infoListener staking
When the app is started when a game launches:
- the background window calls setListeners
- the in game window or second screen calls setListeners
In the function the removeListener call is supposed to remove the events from background an replace them by the one from ingame.
This does not work because the object sent to the function are not the same.
This results as events being triggered twice everytime. For example the moving event in rainboww six pops twice on each move.
The fix here store the object of the last added event to be able to remove it afterward.